cleaning PR after tests
